# Optimized the calculation of total branch lengths of a clade