This PR improves the performance and customizability of `TreeNode.total_length` (renamed from `descending_branch_length`; old name kept as an alias). Test on a tree with 16k tips:

Entire tree:
- Before: 23.8 ms ± 474 μs
- After: 10.9 ms ± 404 μs

5000 tips
- Before: 56.8 ms ± 2.09 ms
- After: 10.7 ms ± 724 μs

2000 tips
- Before: 38.7 ms ± 809 μs
- After: 4.79 ms ± 118 μs

1000 tips
- Before: 32.5 ms ± 1.09 ms
- After: 2.81 ms ± 47.9 μs

***

Please complete the following checklist:

* [x] I have read the [contribution guidelines](https://scikit.bio/contribute.html).

* [x] I have documented all public-facing changes in the [changelog](https://github.com/scikit-bio/scikit-bio/blob/main/CHANGELOG.md).

* [ ] **This pull request includes code, documentation, or other content derived from external source(s).** If this is the case, ensure the external source's license is compatible with scikit-bio's [license](https://github.com/scikit-bio/scikit-bio/blob/main/LICENSE.txt). Include the license in the `licenses` directory and add a comment in the code giving proper attribution. Ensure any other requirements set forth by the license and/or author are satisfied.

  - **It is your responsibility to disclose** code, documentation, or other content derived from external source(s). If you have questions about whether something can be included in the project or how to give proper attribution, include those questions in your pull request and a reviewer will assist you.

* [x] This pull request does not include code, documentation, or other content derived from external source(s).

Note: [This document](https://scikit.bio/devdoc/review.html) may also be helpful to see some of the things code reviewers will be verifying when reviewing your pull request.